Añadido:  - Delete Task.   - Boton Borrar pagina Cesta.

+++ src/main/java/com/smc/casa/web/TaskController.java
@@ -69,10 +69,15 @@ public class TaskController {
     public RedirectView deleteTask(Model model, @PathVariable("id") long id) {
         try {
             Task taskToDelete = service.findById(id);
-            Task deletedTask = service.deleteOne(taskToDelete, id);
-            List<Task> taskList = service.findAll();
-            model.addAttribute("taskList", taskList);
-            return new RedirectView("/indexCestaCompra");
+            if(taskToDelete.isMarked()) {
+                Task deletedTask = service.deleteOne(taskToDelete, id);
+                List<Task> taskList = service.findAll();
+                model.addAttribute("taskList", taskList);
+                return new RedirectView("/indexCestaCompra");
+            }
+            else {
+                return new RedirectView("/indexCestaCompra");
+            }
         } catch (Exception e) {
             System.out.println("Error " + e.getMessage());
             return new RedirectView("error");
